Move schemas.py into movies.py under schemas dir
